Don't override the default router
In order to use the DynamicRouter from the CMF bundle, replace
router.default in cmf_routing.chain.routers_by_id with
jms_i18n_routing.router. This way, both internationalised routes and
dynamic routes will work.

Of course, the way in which this happens means that there's no
localisation on dynamic routes, which may or may not work in every case.